+++ lpbc-label.h
@@ -2,6 +2,6 @@
 #define __LPBC_LABEL_H__
 
 GtkWidget * lpbc_label_new (void);
-void lpbc_label_set_brightness(GtkWidget *label, long brightness);
+void lpbc_label_set_brightness(GtkWidget *label, int brightness);
 
 #endif /* __LPBC_LABEL_H__ */
